Add proxyquire to test globals and add eslintrc
Proxyquire is currently used in a lot of test files so makes
more sense to move it to mocha's globals.

Alongside that this change makes use of the `root` global
variable to reference files to be tested and switches to the new
eslint setup which allows rules and globals to be overridden per
folder.