Allow objects to be filtered by removeFalsey filter
Nunjucks allows to use `length` filter to check if object is empty. This allows us to check objects have values before
passing them to components. If objects has `undefined` or `null` values it might be considered not worthy of showing. Allowing
to filter out those values gives ability to only show values that are explicitly set.