Bring breadcrumb dependency into the app
This is because we want to change how the method is called.

The in-app middleware attaches it to the response object and allows
it to be chained with methods like `req.render()`.

It also adds support for a `title()` method on the response object
which means you can set a different title for the current page.