Allow multiple choice field value to be set using consistent property
This uses the `value` property to set the selected option of a
multiple choice field so that it matches the way the value is set
on text fields.

This means that using a form builder is feasible as it can loop over
a list of fields and pass the same properties to the macros.