Refactor local nav middleware to use req.baseUrl
This changes the local nav middleware so that it uses the baseUrl
sent with the request to build the path rather than using the
current url and relative paths in the local nav definition object.

It also adds a shared redirect middleware that will redirect to the
first local nav item.