Fix flash messages so they are not called on redirects
Flash messages were being set on `res.locals` in a global middleware
that was being called on every request. This meant that the
`req.flash()` method was being called on redirects and clearing the
flash messages so they were not being displayed in a view after a
redirect.

This changes removes the global middleware and creates a method that
can be called from templates. This call is then made in place of the
fallback for messages in the local header.