Simplify creation of generic macro configurations
Any field can be converted to a macro configuration used by applyMacro
global function.

A list of fields (which can be either array of fields with name property
or an object with name as key) can be given to transformFieldsObjectToMacrosObject
global transformer which results in the array of macro configurations with
macro name as key and macro props as value.

Second argument to the transformer contains shared properties as well as
macro override. If `macroName` property doesn’t exist on individual field
`macroName` will be looked up in sharedProps.