Fixed issue with multiple choice macro
Multiple choice and form group macro had an issue caused by properties
being mutated instead of copied with changes.

Form macros were mutating properties when they needed to pass a
copy of properties that had been modified. The side effect of
this behaviour was that one macro would changes properties in another
that needed it. In the case of the form group macro it would call
the select function and change the class properties, but then this
meant the groupClass property didn't work.

Created a test that was broken before the fix, and proves the fix works