Allow entity macro titles without links and override metalist modifiers
Change the entity macro so that it only makes the entity title a link
if an id is passed to it, otherwise make it a plain title.

Add an optional parameter to the entity macro 'contentMetaModifier' to
allow an alternative modifier to applied to the metalist macro call
to allow alernative layouts to be requested when generating entity
collections.