Always cast assignee time to an array
The current code hangs if there is only 1 assignee and the value for
`assignee_time` is then a string rather than an array if there were
multiple values on the step.

This ensures that the value is always what is expected, an array, and
therefore doesn't cause the page to hang.