Fix strong override
There is an override for the strong tag which dictates the font
size as well as the font weight. This can be problematic in other
areas so it has been switched to only override font-weight.

It has also been moved to the shame file so that we can address it as
part of that clean-up.