Fix spacing of local header actions when messages appear
There was no specific container around the content and actions which
caused strange spacing issues when the messages component appeared
within the header.

This wraps a consistent container around the content and only adds
the actions container if actions are present. It allows the spacing
to be more consistent when messages appear.