Fixed show investment details errors
Previously if a user saved an investment details form with a single contact or business activity, and that form had an error, then the form redraw would show multiple contacts and activities, all empty. This was caused because the form expects contacts and activities to be an array and if the form submits a single value then that value was passed back to the form to redraw.

The fix for now is to force the contact and activities fields to be arrays when the form is posted.