Remove imports causing stack overflow
Removes the 2 import statements in multiple choice field that caused stack overflows to be show in company, contact and investment details screens for the first few renders.

The import statements seem to fill the stack as nunjucks appears to have limited space to import functions. Subsequence renders work on so this may be due to use of a compiled copy of the template code.

Adjusted the unit tests which required the imports to use select fields instead as are the only children types used.