New middleware to search for limited company collection
Created a new limited company collection middleware so that this can be extracted from the part 2 controller to follow other collection patterns.

Middleware gets the companies house search result, calls the transformer and then modifies them to send the use to the final add screen if a result is selected.