refactor add company router and controllers
Refactored the controllers for add company to follow the collection patterns and the pattern having a post handler fail through to the form render.

Split the add company step 2 into middleware to get search results and a controller to simply render the result.

Removed hidden fields and lookups used to get the company type in step 2 because the current flow dictates that this screen is only used to search for limited companies.