Use the correct method to reset the value of conditional subfields
When resetting the value for subfields this method was setting
the value to an empty string and setting checked to false.

This caused an issue for radio or checkbox elements as they were having
the value they needed removed. So when the form was submitted with
their value it was empty.