When a quote preview exists don't overwrite with quote result
As the middleware is chained it would previously overwrite the quote
preview object with the quote response.

This caused an issue when a previous quote had been cancelled and then
a user navigated to generate a new preview. It would use the values
from the cancelled quote rather than the new quote preview.

This change ensures that if a quote exists on the locals object then
it will not try and get the previous quote values.