Fix error in highlight filter when term contains regex
If a search term is surrounded by * or contains ? or + it breaks the regex builder within the highlight filter. This change removes regex characters from the search term before using it.

Also adds a try/catch to make the function more rubust so a page does not fail if highlight does.