Only apply bold style to non-muted message components
Previously bold was set on all message components and then overridden
to normal weight on muted message types.

This caused some problems with elements like strong as they were set
to a normal font weight.

This change removes the bold on all and only applies it to the types
necessary.