Ensure assignee time is always processed as an array
If there was only one assignee the value for assignee_time would be
a string rather than an array of strings.

This meant that if an invalid value was returned the string would be
split when played back.

This change ensures that it is always processed as an array.

The same applies for actual assignee time when completing an order.