Add getActiveEvents to event repos
Getting events should happen in the event repository. This new method gets the active events using the search API and then filtering out any results with a disabled on value. The current event can be passed into it for when the user is editing a record that has selected a disabled event, so it still appears in the drop down.