Always show uk company in invest requirements
Currently the investment details screen does not show the uk company
if there are no other requirements details.

This change splits the decision to show the requirements table from
the decision to show the add requirements message.

In addition the logic for this decision, and the removal of empty
values, has moved into the controller so different combinations can
be tested.

The result is if there is just a uk company in the requirements
section, it is shown with the add button/message, if there are more
details the edit button is correctly shown.

'fix/DH-1284_investment_show_company' on '84ef717c'.
test/unit/apps/investment-projects/controllers/details.test.js