Improve disabled on filter to support date
Improves the disabled on filter so that rather than simply test for the presence of  a value it filters items marked as disabled after a given date or the current date.

By using the filter in this manner the front end can build a list of items that were enabled at a given point of time, such as the creation date and time of a record being created.