Filter options, fix address bug & refactor company edit
This change refactors the company edit middleware and addresses a number of issues:
- Filter out disabled options
- Fix bug not setting the country in some cases for uk registered and trading address
- business type was being set in 2 places, this has been changed to just 1
- populateform was long and difficult to understand

Populate form has been simplified to concentrate on building the form state and the job of getting filtered options for the page has been move into it's own function

The job of setting the country for addresses has moved to the form post handler as this is the correct place to handle it. The previous method was causing a big which in some cases stopped using being able to edit a company.

The tests for the middleware have been re-written in a context style to test the differenent ways the controller is called, and repalce stub with Nock to holistically test all the components involved to improve test coverage.