Filter out undefined items returned by item transformers
Previously if transformApiResponseToCollection called an item transformer and that transformer failed it would cause the items collection to contain an undefined item. This caused an issue when other transformers, or users of the transformed data, expected the data to be populated.

The issue was discovered when the contact list item transformer failed for partial names and the following transformer threw an error