Update investment team to use PUT instead of DELETE/POST
There is a bug when LEP users delete all team members and then try to re-add because they may delete themselves and leave the record without team members and not able to update it.

Switching to PUT moves this all up to the server which does it all in one go, so the data it updated though they may still accidently remove themselves.