Switch investment team update to use new `PUT` method in API
Previously the front end had to remove all team members and then re-add them to allow people to change team members, this causes a bug if the user is a LEP as their permission to update the project is deleted and then when it tries to re-add the record they cannot, the result is all team members get deleted.

The PUT method lets the backend do the change in one transaction which fixes the issue.