Simplify routing for editing team members
The investment team members middleware and controllers were overly complex for no good reason. This simpler routing follows other examples in that it populates a form from data in the API or it handles a form post which flows through to the render controller if there is an error to display.