Re-factored edit team middleware
Refactored the middleware for handling the team edit page to have 2 main handlers:

populateTeamEditForm: Takes an investment data entity and generates the form data for it to be displayed in an edit form

postTeamEdit: Handle posting team data and tries to save it. if it works it redirects, otherwise flows through to allow the errors to be rendered.

Transforming team data between different formats is split out into seperate functions to keep things easy to read and follow.

The post handler has been updated to handle the new error format returned by the PUT method int he API which illustrates which team member has an error if something is wrong.