Fix issue handling empty expected land dates
Previous if no land date was provided an empty string was sent to the server which parsed it and turned it into a date in Jan 1970, if there is no expected land date the server expects a null value. Updated the form handler to accomodate this, and this makes it consistent with actual land date.