Prevent getPageTitle error not rendering the layout
A helper function is used to get the page title for the current page.

This template is also used for the error is triggered and the error
page is rendered before the locals middleware has been called and had
time to set this helper function.

This change as some protection around calling this function so that it
will fail gracefully rather than render only this error.