Move saving of OMIS data to saveValues method
This change splits out what the current successHandler method is
responsible for. It was previously handling saving of the data to the
API and redirecting on success.

Looking again at the lifecycle it makes more sense that the saving of
data is taken care of in the saveValues method then the next method
is called to move to the successHandler.

This means the successHandler is then only responsible for clearing of
the session data and redirecting to the next URL. It means repetition
in those code when a custom method for saving is required can be
removed.