Fix fragility when testing collection header
Previously when testing the collection header the tests assumed that there would be 0 or > 1 results. If the result actually had 1 result the heading wouldn't match the test text. This is because the heading uses the pluralise function to decide if it should say 'company' or 'companies'.

This change rewords the feature files to always use the singular form and pluralises it using the same nunjucks filter, if required in the step definition.