Move field processing to process() instead of saveValues()
In the parent OMIS Form Controller there was some logic within the
`saveValues` method which converted repeatable fields into an array
when they were returned as a string.

With the recent changes to the Edit Controller this method was being
overridden and this was no longer happening.

Thinking about what was being done in this method it makes more sense
to handle this processing in the `process()` lifecycle method
rather than in the `saveValues()` method.