Remove collection navigation step definition
This step definition does something very similar to the new common
navigate to page step definition.

It is being removed and its uses replaced with the common step.

A new step has been created to store the value of the count which was
previously being done on every navigate too, however this value is
only used in certain scenarios where this new step is then called.