Remove duplciate company details tests
I have removed the comapnies/details.feature tests for two reasons:

1) They are relying on the company being in the same state as the
fixtures which I am seeing failures around otehr tests changing the
company and then these tests expecting the company to be in the same
state as the fictures.
2) We are already testing details in the companies/save.feature file so
this is a duplication of that code.

I have also moved over the account manager work over to the companies/save.feature
and have updated the scenario text.