Fix rounding errors during OMIS reconciliation
It appears the previous fix hadn't caught all possibilities.

It was using `parseInt` which would round down rather than to the
nearest whole number.

This caused the API to return a possible value mismatch on the
total cost of the order.

Using `Math.round` ensures it will round to the closest integer and
avoid the API returning a mismatch in values.