Call breadcrumb method in form wizard render function
The breadcrumb method was being called in the configure method which
is one of the first methods. If the heading form option was changed
dynamically within a set it would not pick up this change.

This change method ensures it is called later within the lifecycle just
before the template render is called so changes to the value are
still rendered correctly.