Rename client details step within OMIS create journey
Currently we get the company ID from the URL and the first step is
to select the client details. The client is however, the company.

This change renames the client details step to be contact specific
to make way for the changes coming to add selecting the company as
a preceeding step.