Use HMPO form wizard middleware to set company local
This uses the concept of middleware for the form wizard to set a
company locals property that will be available within the views if
there is a company ID set within the session model.

The company is used across a few views within this journey and is
currently set using the shared middleware from the OMIS app to get
it from a URL param.

The new journey will no longer contain an ID within the URL so needs
an alternative way to be set.