Display a variation of the company step when it is set
When a company has already been set and the user visits this step
again, either by editing from the summary or by going back from the
following step, we don't want to show the empty search or same results.

So instead we play back the current company with an option to change it
and bring up the empty search again.