Update contact step to use form wizard middleware
This step does some changes to the contact options and the heading
based on the company that exists in the session.

This company ID previously came from the URL but the new journey
doesn't add the company ID as a URL param.

This step therefore needs to get that value from the session and
the session is not available during the `configure` method.

This change makes use of the middleware methods that are called within
the lifecycle and adds this logic to a place where the session model
now exists.