Redirect to redirect property in OMIS controller error handler
If the error that is sent to the form controller contains a redirect
property the error handler should redirect there rather than calling
the parent error handler.

This change also negates the need to specifically handle the
`MISSING_PREREQ` error as within the form wizard this URL is already
decided and sent with the error object.