Fix bug with item modifier leakage
Meta list was using assign to modify properties sent
to a meta item, this meant if a modifier was applied to
an item it would change the original data. This showded
up in the component gallery when styles from one example
would leak to another.

Changed the way modifiers are applied for meta-lists and
meta-list-item to be consistent