Ensure correct prop value is passed to select box macro
The top-level multiple choice macro supports a prop to add a class
to the lower field.

This was being referenced on its own and not as a property on the
props object so the value was never picked up.

This change makes sure the lower macro receives the correct value.