Increase dockerize leeloo timout to 5 mins
We are seeing a few timeouts on ES when leeloo is being spun up for
CirclCi. This work increases the timeout to 5 mins to give it a little
more breathing space.
Its worth keeping an eye on this as when more migrations to the BE in
Postgres are added the time it will take for leeloo to spin up on
CircleCi will increase.