Push new history before response returns
Previously the URL and state are only updated once the response
is returned on the xhr call. This leads to a problem when the call
errors and you refresh the page and the selected filters and state
are lost.

This change adds a simple push to the browser history before the call
and this url is then updated on a successful xhr request.