Return from methods in AT tests
The fact that this methods were not having this returned in them meant
that there were a number of race conditions where the tests moved on to
the next scenario before this one was finished. This meant state was not
updated and the tests failed. Adding in the return means that the tests
will wait until the method has finished and returned.

Also a random extra callback has been removed.