Add vue js and .vue support to build process
This change introduces the ability to scan the page
for vue js component and activate them

The implemtation uses the .vue file format over
embedding templates into nunjucks so components
can be nested, properties can be specified
and it dramatically improves page display
when using compiled templates.

This will allow a component to be embeded
in a page and if the user does not have js support
the component will simply not appear.