Remove fixtures from the world object
Fixtures are currently set on the world object and accessed using
`this` within test files.

We don't really need to do this as the fixtures are a file exported
within the same parent directory. Files which use fixtures can just
load the fixtures directly rather than clogging up the world object
with all the fixtures.

This change removes the fixtures from the world and only loads the
fixtures into files that actually use them.