Use direction navigation in company tests
With the new idea of page objects we can navigate directly to a
particular sub-page within an entity.

This means we don't have to navigate to the details page and click
a nav item which adds extra time to the tests to do all the page
loads and page interactions.

This removes those extra steps within the company tests.