Do all audit checks in one scenario
Currently we run a different scenario to check that an audit
item contains the correct data.

This feels quite intensive as for some scenarios it includes searching
for a company record, navigating to the contacts screen, adding a new
contact, navigating to that contact, navigating to the audit tab
and then checking the contents of the last audit item.

We can save lots of requests by only doing the checks once.