Add support for more Nightwatch commands and assertions
This imports a new library to add further useful commands and
assertions to our acceptance testing framework, Nightwatch.

One particularly useful assertions is being able to check a URL using
a regex rather than checking for it matching or containing a string.

This will allow us a better way to check the correct page object
for page objects that expect a UUID in the URL.