Add step definition to select form values on a page object
This step definition allows a value to be selected from a dropdown
on a particular page object.

It makes use of the state object to store that value so that it can be
checked against in following tests if necessary.

It saves the value using the page object path and the element
reference.

For example, `omis.create.Contact.@contactField`