Add step definition to check current page
This step definition allows you to check you are on the desired
page object by comparing the current URL to the expected URL of that
particular page object.

If the page object contains a prop for matching the URL (`urlMatch`)
it will check using a regex rather than looking for the URL to contain
the value of `Page.url()`.