Create basic test for OMIS creation journey
This change makes use of the page objects and shared step definitions
to run through the simplest order creation journey for when a company
is know and the ID is passed to the creation starting point.

It includes the addition of a custom step definition to save the
value of the company's default sector.