Rename acceptance test page objects
This renames the page objects to be lowercase. The reason for this is
so that we don't have different cases when references page objects
within feature files.

Currently we have examples like `companies.Contacts`. This rename
makes the page objects and folders more consistent so they are slightly
easier to read within step definitions.

If we want to go one more step to remove the dots inbetween too we
won't have to worry about changing case if they're all the same case.