Remove vat status step from quote errors
The vat status step is a nested step used when changing the billing
address country and also contains the `vat_status` field.

Adding this step caused it to be caught by the quote middleware and
it to show the same error twice with a link to 2 different steps.

As the vat status step is hidden within a journey we want to remove
it from this quote logic.

This change also contains updates to the test to catch this change.