Change EU VAT number example
Previously the error message example used Austria including the
country code. Austria's VAT number format always starts with the
letter `U` and when combined with the country code of `AT` it
confused some users who thought it was suggesting the country code
should be made up of 3 characters.

This change uses Italy as an example which has a more easily
understandable format.