Amend error message on estimated hours
When there are no advisers in the market and a user tries to add
estimated hours we show a message box with why this cannot be done
yet and what needs to be done.

Previously this was using the info variation which does not signify
as well that something has to be done first.

This now uses the pattern from the quote preview journey.