Don't rely on fieldset modifier for field error styles
We currently rely on a `subfield` modifier being applied to the
fieldset component to adjust styles when a fieldset is being displayed
as a subfield and contains errors.

This is the only use of modifiers being used by the fieldset component
and we would like to remove all those extra classes from the form group
component if possible.

This changes moves those style overrides so they are no all contained
in the form group style component.