Change playback of client email address when sending quote
This change amends how we display what email address the quote
email will be sent to for the client.

We previously used a _What happens next_ heading and some text within
a list but this was often missed during research. We also displayed
the email address within the action button but in some cases of a long
email address this would break the layout of the page and also loses
the consistency in form actions.

This change uses a pattern tested on GOV.UK Pay to replay the email in
a larger typography before the user submits the form.

It also uses a consistent language for the button rather than changing
the value.