Add support for hidden labels in forms and filter forms
The addition of the adviser filters to the interactions
collection means there are two controls in the form that
filter adviser, to avoid repitition of the same label
or legend this feature allows a label to be
visually hidden.

When used in filter forms a small adjustment is required
to padding to keep the layout consistant.