Consistently refer to the market the same way
In other instances of the service we use a new, more explanatory term
and the old, legacy term in brackets. For market we have been using the
newer term in brackets.

This change uses country first, then market in brackets when we feel
it is necessary to clarify we're referring to the market.

It also switches the heading and field label to be consistent with the
rest of the create journey.