Handle no contacts on create journey better
Currently when there are no contacts for a company we still show
the form field which simply has no items.

This doesn't provide the best experience so this piece of logic
hides the form field when there are no contacts and provides an
alternative, more tailored message to this situation to create a
new contact.