Require verification before completing order
One of the current requirements is that the work must be sent to the
contact before an order is completed.

Previously this was a message but whilst the system doesn't support
the uploading and sending of the documents a change is added to
require the person completing the order to verify that the work
has been sent to the contact.