Tweak block entity item style
This removes the underline on the block style link. The underline
suggests only the heading is clickable whereas in this variation
we are making the whole block area clickable.

It also reduces the padding on the top and bottom to cater for line
height so that visually the text is more aligned.