Use separate prop for hiding error message on form group
Currently if you use the `isLabelHidden` prop it also hides the error
message. There are some cases where we don't want to display the
whole label but still want to show the error message and border
when there is a problem with the field.

This introduces a separate prop for hiding the error message as well
as the label.