DH-1627 Clear filters functionality doesn't apply for typeahead filter
Because typeahed filters have a different behaviour than normal filters we had to find a way of clearign them before the submit to the server.
We were able to clear the filter using normal vue behaviour but this was creating a race condition with the global filters submit,
teherefore decided for this approach, by removing the filter field from the clear-filters submit.

DH-1627 refactored for IE and iOS support of js remove

DH-1627 changed forEach statement to work for IE and firefox