Rename collection macros
The naming of collection related nunjucks macros meant
that we had 4 different files named collection.njk and it was
not clear what some of the collection files did from their
name.

This refactor renames the different macros and pages related
to collections to make it clear what their purpose is, and also
splits out the collection list portion to it's own macro.

The naming now follows the structure described in the wiki
https://github.com/uktrade/data-hub-frontend/wiki/Collections