Remove actions support and introduce add button option to collections
The actions html feature in collections is only used for the add button
so it is simpler to add a 'add button' feature to collections, requiring
a label and url. Any changes to where the button is placed or how it
acts or is styled can then be maintained in one place and some templates
can be removed altogether.