Create a new function to get interaction options
Previous interaction options was a piece of middleware and it always
got all interaction related options, even when not needed. Also the
middleware was only used by one thing so it makes more sense
to make it a function called from a controller.

The code is now not only smarter and easier to read but also
uses the new context support in the options function.