Adds new tests for edit interaction
Adds BDD style tests that are more integration like than before. The
tests test adding and editing an interaction from the request all
the way through to network calls to the backend.

The first tests focus on a regular interaction type, a follow up
will add tests for service deliveries.