Fix bug picking services for invstment interactions
The previous code assumed that when an interaction was
to be added to an investment project the router would
pass the investmentId parameter, but for some reason when
using sub routes the parameters from the parent route
are marked undefined.

This change uses the investment data loaded by previous
middleware or the investment project id in the interaction
to judge if it's a investment interaction or not.

Also fixed the tests that missed this.