Update babel config to use babel-preset-env
Babel reccomends that users move to a new way of
specifying modules and plugins for transpiling code.

By moving to babel-preset-env you can specify the
target browsers and it handles analysing the code
and fetching required modules. This also allows
new javascript syntax to be adopted without changing
the babel config.