Fetch company subsidiaries via v3 api call
Previously the list of company subsidiaries was fetched using the search
API. There is a slight delay populating search so this change uses the
v3 company API and a filter to get the same information from the database.

Refactored subsidiaries view to populate all properties in the controller
instead of the view

Added tests to confirm new functionality.