DHI-9 Investments SPI - SPI3 Feature
For investment projects it's required a measuring sysem called SPI - Service Performance Indicator.
The process itself has multiple SPI's and we're implementing in DataHub from SPI1 to SPI5.
In this feature we focused on SPI3 which represent investment projects propositions. Propositions should be a subcategory
of Investments projects. In this commit I created the core functionalities for SPI3:
  * Create a proposition
  * List all propositions
  * See the details of a proposition
  * Abandon a proposition
  * Complete a proposition - NB: This feature requires document upload, which is not in scope for this release, so we had to implement a simple textarea that will contain a sharepoint link which is added manually by the user.