Fix issue sorting attendees
If a user sorted event attendees, refresh the page then try another sort it would fail to sort. If a user sorted twice in a row the same thing happened.

The issue is the auto submit script posts the form to the form action value
or the page url if there is none. As the page url has the sort order in it
already both values were sent, which causes it to fail.

The fix simply specifies an action url for the sort form so the values are
always posted to the correct url.