Shown known details for attendees without contact
For some reason some event attendance has been
recorded without a contact, only the company. This
change modifies the attendee transformers to
inform the user, and change the title link to link
to the attendance so they can update it.

The transformation for company has also been
made more robust, though the database scheme
always requires a company.