https://trello.com/c/ZnsFf3u9/171-event-attendees-double-clicking-adds-the-same-attendee-twice
Some users still double click on links and this creates an unwanted behaviour, i.e. multiple entries.
I've done some discovery on what's the best way to aproach this, and becuase in the node layer it creates multiple `threads`
you can't really control the state of multiple requests unless you create an observer pattern in the authorised-request file.

Knowing that the authorised-request file is being used with almos all request, this pattern is not fit for purpose so I decided to solve it
very simply in the client by preventing any double or multi click actions on a link with a predefined class, thus preserving the native browser behaviour.