Rename parent to global headquarters to avoid disambiguation
A parent/child relationship between companies assumes that
there can be more levels (e.g. a child company with its own child).

The hierarchy we have at the moment is Global HQ and subsidiaries
so it's better to avoid the word 'parent' and use
global headquarters instead.

Also note that a Global HQ might not be the direct parent of a
subsidiary in real life so using 'parent' is not 100% correct.
We don't want to map the real world relationships between
companies as it's too unnecessarily complex.