Documents module cleanup
Initial version of Documents module was using a constructor pattern and most of the logic was in a single file.
In this commit we untangled the logic in separate files, easier to go through, and also using the project common patterns.
The costructor pattern was changed to a functional one. Also moved the information from the class scope `self` into `response.locals.documents` thus clearing the arguments list carried through the functional flow