Propositions Documents Feature
For Propositions there is a user need to have a documents upload feature. We're using the `documents` module recently created,
and we're instantiating it in the `router.sub-app` section of Propositions by calling the `postUpload` function and binding to it the api configuration
i.e. `url:{app:'investment', subapp'permision'}`. This config is necessary to build the api urls for the requests.

Complete Proposition Journey

Because now we're using documents upload we no longer need the work-around view with a text input for a sharepoint link,
so we've removed it.